+++ Mimus/Source/Verification/VerificationHandler.swift
@@ -21,6 +21,12 @@ internal class VerificationHandler {
                     matchCount: matchedResults.count,
                     differentArgumentsMatchCount: mismatchedArgumentsResults.count,
                     testLocation: testLocation)
+        case .atMost(let count):
+            assertAtMost(callIdentifier: callIdentifier,
+                    times: count,
+                    matchCount: matchedResults.count,
+                    differentArgumentsMatchCount: mismatchedArgumentsResults.count,
+                    testLocation: testLocation)
         case .times(let count):
             assert(callIdentifier: callIdentifier,
                     times: count,
@@ -42,14 +48,29 @@ internal class VerificationHandler {
                 line: testLocation.line)
     }
 
+    private func assertAtMost(callIdentifier: String, times: Int, matchCount: Int, differentArgumentsMatchCount: Int, testLocation: TestLocation) {
+        guard matchCount > times else {
+            return
+        }
+
+        var message = "No call with identifier \(callIdentifier) was captured"
+        if matchCount > 0 {
+            message = "Call with identifier \(callIdentifier) was recorded \(matchCount) times, but expected at most \(times)"
+        }
+
+        XCTFail(message, file: testLocation.file, line: testLocation.line)
+    }
+
     private func assertAtLeast(callIdentifier: String, times: Int, matchCount: Int, differentArgumentsMatchCount: Int, testLocation: TestLocation) {
         guard matchCount < times else {
             return
         }
 
         var message = "No call with identifier \(callIdentifier) was captured"
-        if differentArgumentsMatchCount > 0 {
-            message = "Call with identifier \(callIdentifier) was recorded, but arguments didn't match"
+        if matchCount > 0 {
+            message = "Call with identifier \(callIdentifier) was recorded \(matchCount) times, but expected at least \(times)"
+        } else if differentArgumentsMatchCount > 0 {
+            message = "Call with identifier \(callIdentifier) was recorded \(differentArgumentsMatchCount) times, but arguments didn't match"
         }
 
         XCTFail(message, file: testLocation.file, line: testLocation.line)
@@ -64,7 +85,7 @@ internal class VerificationHandler {
         if matchCount > 0 {
             message = "Call with identifier was recorded \(matchCount) times, but expected \(times)"
         } else if differentArgumentsMatchCount > 0 {
-            message = "Call with identifier \(callIdentifier) was recorded, but arguments didn't match"
+            message = "Call with identifier \(callIdentifier) was recorded \(differentArgumentsMatchCount) times, but arguments didn't match"
         }
 
         XCTFail(message, file: testLocation.file, line: testLocation.line)