+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Next (1.1.2)
 - Added out of the box comparison support for `UInt`
 - Fixed issues with comparing dictionary or array with a different yielded an incorrect message
+- Add '.atMost' VerificationMode
+- Make failure messages more detailed 
 
 
 # 1.1.1